+++ types/protocol.d.ts
@@ -12093,6 +12093,21 @@ export namespace Protocol {
             resource: LoadNetworkResourcePageResult;
         }
 
+        export interface SetCookieControlsRequest {
+            /**
+             * Whether 3pc restriction is enabled.
+             */
+            enableThirdPartyCookieRestriction: boolean;
+            /**
+             * Whether 3pc grace period exception should be enabled; false by default.
+             */
+            disableThirdPartyCookieMetadata: boolean;
+            /**
+             * Whether 3pc heuristics exceptions should be enabled; false by default.
+             */
+            disableThirdPartyCookieHeuristics: boolean;
+        }
+
         /**
          * Fired when data chunk was received over the network.
          */